record crawler stats

+++ data_crawler/crawler.js
@@ -73,6 +73,9 @@ const crawlDoi = async (paperid) => {
   if (!res.ok) {
     return [{}, res.status]
   }
+  if (res.status === 200) {
+    stat.semantic200Cnt += 1
+  }
   const json = await res.json()
   return [{
     title: json.title,
@@ -83,7 +86,7 @@ const crawlDoi = async (paperid) => {
 /**
  * 
  * @param {string} title 
- * @return {[doi, number ]}
+ * @return {[string | null, number ]} [semanticScholarId, statusCode]
  */
 const crawlIdByTitle = async (title) => {
   const endp = `https://api.semanticscholar.org/graph/v1/paper/search?query=${title}&limit=1`
@@ -91,9 +94,12 @@ const crawlIdByTitle = async (title) => {
   if (!res.ok) {
     return [{}, res.status]
   }
+  if (res.status === 200) {
+    stat.semantic200Cnt += 1
+  }
   const json = await res.json()
   if (json.total === 0) {
-    return [{ title, abstract: null }, res.status]
+    return [null, res.status]
   }
 
   return [json.data[0].paperId, res.status]
@@ -122,7 +128,7 @@ const exists = (path) => fsSync.existsSync(path)
  * @param {string} venue 
  * @param {string} title 
  * @param {string} abstract 
- * @returns {Promise<Error | null>}
+ * @returns {Promise<void>}
  */
 const save = async (year, venue, title, abstract) => {
   // const fs = {
@@ -133,16 +139,26 @@ const save = async (year, venue, title, abstract) => {
   //   mkdir: nop,
   //   writeFile: nop,
   // }
-  const filepath = toFilePath(year, venue, title)
-  let err = await fs.mkdir(`data/${year}/${venue}`, { recursive: true })
-  if (err !== undefined) { return err }
+  if (abstract === "") {
+    stat.noAbstractCnt += 1
+  }
+  try {
+    const filepath = toFilePath(year, venue, title)
+    let err = await fs.mkdir(`data/${year}/${venue}`, { recursive: true })
+    if (err !== undefined) { throw err }
 
-  if (!exists(filepath) || isEmpty(filepath)) {
-    err = await fs.writeFile(filepath, abstract)
-    if (err !== undefined) { return err }
+    if (!exists(filepath) || (isEmpty(filepath) && abstract !== "")) {
+      err = await fs.writeFile(filepath, abstract)
+      if (err !== undefined) { throw err }
+      stat.saved += 1
+    } else {
+      stat.notSaved += 1
+    }
+  } catch (e) {
+    stat.notSaved += 1
+    stat.failed.push({ title, doi: "unknown", error: e.toString()})
   }
 
-  return null
 }
 
 /**
@@ -151,14 +167,30 @@ const save = async (year, venue, title, abstract) => {
  * @returns 
  */
 const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, msec));
+const stat = {
+  noAbstractCnt: 0,
+  titleMismtachCnt: 0,
+  currentVenue: "",
+  semantic200Cnt: 0,
+  reqCnt: 0,
+  failed: [],
+  saved: 0,
+  notSaved: 0,
+};
 
 (async () => {
-  let failed = []
   let dblpQ = dblp.map(e => e)
   let semanticsScolarQ = []
-  let succeed = 0
+  const start = process.hrtime();
+
+  const elapsed = () => +process.hrtime(start)[0]
+
+  const printStat = () => {
+    const perSec = (metric) => (metric / elapsed()).toFixed(3)
+    console.log(`${stat.currentVenue} dblpQ.len ${dblpQ.length} scholarQ.len ${semanticsScolarQ.length} Ok ${perSec(stat.semantic200Cnt)}/s saved ${stat.saved} (${perSec(stat.saved)}/s) notSaved ${stat.notSaved} (${perSec(stat.notSaved)}/s) stat.failed ${stat.failed.length} requeue ${stat.reqCnt} (${perSec(stat.reqCnt)}/s) titleMismtachCnt ${stat.titleMismtachCnt} (${perSec(stat.titleMismtachCnt)}/s) noabs ${stat.noAbstractCnt} (${perSec(stat.noAbstractCnt)})`)
+  }
   
-  while (199 < dblpQ.length) {
+  while (222 < dblpQ.length) {
     dblpQ.pop()
   }
 
@@ -170,6 +202,7 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
     // confernce := <- dblpQ
     if (dblpQ.length === 0) { return; }
     const conference = dblpQ.pop()
+    stat.currentVenue = `${conference.year}/${conference.name}`
 
     const [papers, errs] = await crawlDblp(conference.url)
     if (0 < errs.length) {
@@ -199,9 +232,15 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
     // get semantic scholar ID and re-queue
     if (!doi) {
       const [id, statusCode] = await crawlIdByTitle(title)
+      if (id === null) {
+        // we have no abstract
+        await save(year, name, title, "")
+
+        return
+      }
       switch (statusCode) {
       case 200:
-        console.log("200 requeue ", title)
+        stat.reqCnt += 1
         semanticsScolarQ = [
           { year, name, title, doi: id },
           ...semanticsScolarQ
@@ -209,14 +248,14 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
         break;
       case 429:
         // too many requets. requeue
-        console.log("429 requeue ", title)
+        stat.reqCnt += 1
         semanticsScolarQ = [
-          { year, name, title, doi },
+          { year, name, title, doi: id },
           ...semanticsScolarQ
         ]
         break;
       default:
-        failed.push([title, doi])
+        stat.failed.push({title, doi, error: `crawlIdByTitle returned ${statusCode}`})
       }
       return
     }
@@ -225,9 +264,9 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
     switch (statusCode) {
     case 200:
       break;
-    // cannot find paper by DOI. Try semantic scholar ID
     case 404:
-      console.log("404 requeue ", title)
+      // cannot find paper by DOI. Try semantic scholar ID
+      stat.reqCnt += 1
       semanticsScolarQ = [
         { year, name, title, doi: null },
         ...semanticsScolarQ
@@ -236,7 +275,7 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
 
     // Too frequent requests. Just requeue.
     case 429:
-      console.log("429 requeue ", title)
+      stat.reqCnt += 1
       semanticsScolarQ = [
         { year, name, title, doi },
         ...semanticsScolarQ
@@ -244,41 +283,29 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
       return
 
     default:
-      failed.push([title, doi])
+      stat.failed.push({ title, doi, error: `crawlDoi returned ${statusCode}`})
       return
     }
     
-    {
-      try {
-        if (title !== paper.title) {
-          console.error(`title mismatch ${paper.title} -> ${title}`)
-          const filepath = toFilePath(year, name, paper.title)
-          if (exists(filepath)) {
-            const newFilePath = toFilePath(year, name, title)
-            fsSync.renameSync(filepath, newFilePath)
-          }
-        }
-        const err = await save(year, name, title, paper.abstract || "")
-        if (err !== null) {
-          failed.push([title, doi])
-        } else {
-          succeed += 1
-        }
-      } catch (e) {
-        console.error(e)
-        failed.push([title, doi])
+    if (title !== paper.title) {
+      stat.titleMismtachCnt += 1
+      const filepath = toFilePath(year, name, paper.title)
+      if (exists(filepath)) {
+        const newFilePath = toFilePath(year, name, title)
+        fsSync.renameSync(filepath, newFilePath)
       }
     }
+    await save(year, name, title, "")
   }
 
   const dblpWorker = setInterval(runDblp, 3000);
-  const semanticScholarWorker = setInterval(runSemanticScholar, 1050);
+  const semanticScholarWorker = setInterval(runSemanticScholar, 1100);
 
   while (true) {
     await sleep(3000)
 
-    console.log(`dblp ${dblpQ.length} scholar ${semanticsScolarQ.length} succeed ${succeed} failed ${failed.length}`)
-    fsSync.writeFileSync("./data_crawler/failed.json", JSON.stringify(failed, null, 2))
+    printStat()
+    fsSync.writeFileSync("./data_crawler/failed.json", JSON.stringify(stat.failed, null, 2))
     if (dblpQ.length === 0) {
       // 10 sec room
       await sleep(10_000)
@@ -286,7 +313,7 @@ const sleep = async (msec) => await new Promise(resolve => setTimeout(resolve, m
         // all queues are empty
         clearInterval(dblpWorker)
         clearInterval(semanticScholarWorker)
-        console.log(JSON.stringify(failed, null, 2))
+        console.log(JSON.stringify(stat.failed, null, 2))
         process.exit(0)
       }
     }
